Module aws_sdk_snowdevicemanagement::error

Modules

access_denied_exception

See AccessDeniedException

internal_server_exception

See InternalServerException

resource_not_found_exception

See ResourceNotFoundException

service_quota_exceeded_exception

See ServiceQuotaExceededException

throttling_exception

See ThrottlingException

validation_exception

See ValidationException

Structs

AccessDeniedException

You don't have sufficient access to perform this action.

CancelTaskError

CreateTaskError

DescribeDeviceEc2InstancesError

DescribeDeviceError

DescribeExecutionError

DescribeTaskError

InternalServerException

An unexpected error occurred while processing the request.

ListDeviceResourcesError

ListDevicesError

ListExecutionsError

ListTagsForResourceError

ListTasksError

ResourceNotFoundException

The request references a resource that doesn't exist.

ServiceQuotaExceededException

The request would cause a service quota to be exceeded.

TagResourceError

ThrottlingException

The request was denied due to request throttling.

UntagResourceError

ValidationException

The input fails to satisfy the constraints specified by an Amazon Web Services service.

Enums

CancelTaskErrorKind

CreateTaskErrorKind

DescribeDeviceEc2InstancesErrorKind

DescribeDeviceErrorKind

DescribeExecutionErrorKind

DescribeTaskErrorKind

ListDeviceResourcesErrorKind

ListDevicesErrorKind

ListExecutionsErrorKind

ListTagsForResourceErrorKind

ListTasksErrorKind

TagResourceErrorKind

UntagResourceErrorKind