Module aws_sdk_snowdevicemanagement::model

Modules

capacity

See Capacity

cpu_options

See CpuOptions

device_summary

See DeviceSummary

ebs_instance_block_device

See EbsInstanceBlockDevice

execution_summary

See ExecutionSummary

instance

See Instance

instance_block_device_mapping

See InstanceBlockDeviceMapping

instance_state

See InstanceState

instance_summary

See InstanceSummary

physical_network_interface

See PhysicalNetworkInterface

reboot

See Reboot

resource_summary

See ResourceSummary

security_group_identifier

See SecurityGroupIdentifier

software_information

See SoftwareInformation

task_summary

See TaskSummary

unlock

See Unlock

Structs

Capacity

The physical capacity of the Amazon Web Services Snow Family device.

CpuOptions

The options for how a device's CPU is configured.

DeviceSummary

Identifying information about the device.

EbsInstanceBlockDevice

Describes a parameter used to set up an Amazon Elastic Block Store (Amazon EBS) volume

ExecutionSummary

The summary of a task execution on a specified device.

Instance

The description of an

InstanceBlockDeviceMapping

The description of a block device mapping.

InstanceState

The description of the current state of an instance.

InstanceSummary

The details about the instance.

PhysicalNetworkInterface

The details about the physical network interface for the device.

Reboot

A structure used to reboot the device.

ResourceSummary

A summary of a resource available on the device.

SecurityGroupIdentifier

Information about the device's security group.

SoftwareInformation

Information about the software on the device.

TaskSummary

Information about the task assigned to one or many devices.

Unlock

A structure used to unlock a device.

Enums

AttachmentStatus

Command

ExecutionState

InstanceStateName

IpAddressAssignment

PhysicalConnectorType

TaskState

UnlockState