Module aws_sdk_snowdevicemanagement::operation

Structs

CancelTask

Sends a cancel request for a specified task. You can cancel a task only if it's still in a

CreateTask

Instructs one or more devices to start a task, such as unlocking or rebooting.

DescribeDevice

Checks device-specific information, such as the device type, software version, IP

DescribeDeviceEc2Instances

Checks the current state of the Amazon EC2 instances. The output is similar to

DescribeExecution

Checks the status of a remote task running on one or more target devices.

DescribeTask

Checks the metadata for a given task on a device.

ListDeviceResources

Returns a list of the Amazon Web Services resources available for a device. Currently, Amazon EC2 instances are the only supported resource type.

ListDevices

Returns a list of all devices on your Amazon Web Services account that have Amazon Web Services Snow Device Management

ListExecutions

Returns the status of tasks for one or more target devices.

ListTagsForResource

Returns a list of tags for a managed device or task.

ListTasks

Returns a list of tasks that can be filtered by state.

TagResource

Adds or replaces tags on a device or task.

UntagResource

Removes a tag from a device or task.